Move edition methods into edition interactor
Previously the Edition class had two methods - 'create_next_edition'
and 'resume_discarded' - that were only used by the Interactor to create
a new edition. This inlines the code for these methods in the Interactor
and in doing so fixes a couple of bugs:

   - Stop trying to create a preview for the live edition
   - Fix state assertion preventing resuming a discarded edition

Previously these two Edition methods set a couple of attributes -
'last_edited_by' and 'last_edited_at' - that are common across the other
'assign' methods. This adds a new 'assign_as_edit' method to abstract
and de-duplicate the setting of these attributes.