Allow support users to override access limits
This adds a new permission that overrides the normal 'foridden' and
filtering behaviours for access limited editions, so that users are
still able to provide support with requests associated with these
editions, which they would otherwise be unable to find/view/edit.

While some existing apps use more role-like permissions, such as 'GDS
Admin', the permission here is specific to its functionality, as well as
which users it should be applied to. This helps avoids confusion about
the behaviour and applicability of permissions that are actually roles.

Previously it was only possible to test this behaviour at the feature
level, which would be complex due to the three variables - permission,
limit type, organisations - involved. This extracts the conditional
into a new User#can_access? method in order to fully test the code.