DRY-up requirements specs with a new matcher
Previously our requirements specs contained repetitious code for
checking that different kinds issue could be rendered in the context of
a form and (potentially) on the summary page, which we do by checking
the associated translations. This DRYs-up the code to check issues can
be rendered in form/summary contexts by replacing it with a new matcher
to check the issues are equal by checking their associated messages.

In doing so, this also fixes a lack of coverage for the
ImageRevisionChecker, since we did not previously test that the context
included the 'image_revision' variable, the absence of which would have
caused a routing exception if it were not present in production.