Use scopes instead of direct SQL for edition filters
This moves the rather complex SQL to be abstracted into scopes. I
wrote this code a few days ago for a different reason and then once I
saw the earlier code was producing brakeman warnings.

Since we only just cleared out the brakeman ignore I felt a bit sad
about introducing new ones. I felt that this scope based approach
offered code that was more declarative and thus could be easier
followed.

The scope for `primary_organisation_is` is a bit weird but I think
that's more a product of our expectation that this an array of one item
than anything else. Hopefully we can soon change this.