Prepare ContactsController to support requirements
Previously the insertion of a contact required the routes to be under a
document scope, which is no longer the case. This moves the routes for
contacts out of the document scope and adds an '-embed' suffix to be
consistent with the approach for YouTube videos.

Previously we re-used the 'contacts#index' action to re-render the page
and show the snippet to insert. This splits the 'index' action into
'new' and 'create' actions, again to be consistent with the approach for
video embeds. It would have been painful to try and support requirements
checking with the previous 'index' action.

This also changes the insertion mechanism to copy the approach we use for
video embeds, where, in addition to requirements stuff, the controller
responds with the snippet to insert. As a consequence, the autocomplete
no longer needs to have knowledge of the markdown template for contacts.