Split `edit image` button into two actions
On the image index page the link 'Edit image' has been split into two
actions, one being 'Edit details' which will take the user to the edit
image metadata page and the other being 'Crop image' which as you
probably can guess takes the user to crop the image.

There is now a check on where to redirect the user after they have
cropped an image. If the image is being newly uploaded then they will be
taken to the edit metadata page, if the image crop is being edited then
they will be redirected to the images index page after cropping.

The ordering of these links has also been updated to conform with the
new design iteration.