Split PublishingApiPayload into supporting classes
Previously we had a single PublishingApiPayload class to generate
payloads for scheduling and draft preview. This splits up the class
into a couple of supporting classes for their respective services,
as part of the work to tidy up the 'app/services/' directory.

One potential issue was the location of the 'PUBLISHING_APP' constant,
which is used by three services. It makes sense for the constant to be
defined close to the 'put content' payload, since the other two services
that use this constant are coupled to the value specified here.