Encapsulate normalised image in a new class
Previously we had several instances of code that checked or operated on
different attributes of a normalised image in isolation:

   - The ImageUploadChecker needed the normalised dimensions to check
if the image was big enough
   - The ImageUploadChecker re-opened the image to check it was not
animated
   - The ImageBlobService needed the normalised dimensions to pass them to
the centre crop
   - The ImageBlobService needed the normalised file/io to set it on the
ActiveStorage::Blob

This encapsulates the three related attributes - (normalised) 'width',
'height' and 'file' - into a new 'TempImage' class. Instead of passing
the raw params and the new temp file to the ImageUploadChecker, this
moves the checks that rely on it into the ImageNormaliser, taking the
same approach as the DateParser and the DatetimeParser.