Use before_action to check withdraw permissions
Previously we had a couple of approaches to enforcing that withdraw and
unwithdraw actions can only be done by a user with the 'managing editor'
permission. This refactors the enforcement code to consistently use a
before_action and render the associated error page. Since we only have
two instances of pages for this permission, it's not clear that we want
to de-duplicate the code and the views yet.