Refactor PreviewService into two classes
Previously we had a single PreviewService class with an additional
method to silently/gracefully fail when an error occurs but the creation
of the preview is not the main intent of the user action; showing an
error would be confusing/surprising and introduce unnecessary doubt.

This splits the additional silent-fail method into a new 'Failsafe'
version of the PreviewService, as part of our work to refactor our
services into classes that only have one public method i.e. 'they only
do one thing'.