Make asset cleanup a non-time-sensitive job
Previously we had a DraftAssetCleanupService that worked out which
image/attachment revisions were deleted, compared to the preceding
revision, and updated their assets accordingly in Asset Manager and
in the local DB. Because this service only refers to any preceding
revision, it has to be run every time any change is made to ensure
it always has the opportunity to detect any deletions; this subtle
behaviour makes it brittle and prone to missing deletions in future.

This rewrites the service as a job that operates over all editions
and revisions, which means it no longer needs to be run after every
change. Although we could make the job specific to a document, the
longterm intention is for this to run periodically as a scheduled
job, so adding this behaviour doesn't seem necessary.