Fix lead image workflow bug
A while ago a bug was introduced which meant once you cropped a lead
image the crop updater didn't take into account whether it was lead
or not, this was caused by false being passed to the updater from
the crop interactor. This commit fixes the issue.

I have also made the tests more realistic around editing a 'lead' image
as previously the lead image test was just testing with a normal image.
I've also added a step to ensure the image is still lead after editing.