Refactor ScheduleService to have single method
Previously the ScheduleService had both a 'schedule' and 'reschedule'
method, each differing in the way they constructed a Scheduling object.
This moves the construction logic into the lone callers for each method,
reducing the ScheduleService to a single 'schedule' method that is more
compliant with our intended interface for services.

Note that two of the aspects of the construction are still covered at
the feature test level, despite the code moving into interactors:

   - storing the previous edition state (when unscheduled)
   - storing the intended review status for after publishing