Support contact preview when JavaScript fails
Previously we relied on the contact preview being shown in the context
of a modal where the parent page had the Document ID in the URL, which
precludes support for this feature when the view is rendered without JS,
in isolation. This moves the contact-embed paths back into the document
scope in order to support preview for contacts via the GovspeakDocument
class, which requires an edition (to support other kinds of snippet).