Perform asset cleanup with singular queries
This changes the previous method to do asset cleanup so that it can be
done with a single query per asset type. The previous approach that used
ids was at a scaling risk that would mean the application would
eventually crash once the number of items returned reached a
sufficiently high number.