Clean up assets every 30 minutes
This adds the sidekiq-scheduler gem to run the AssetCleanupJob every 30
minutes. This removes the need to run this job manually whenever a
preview is created.

This also the with_advisory_lock gem to ensure that this job is only
running on one particular machine at a time. To ensure a lock name that
is unique from other GOV.UK applications I prefixed the lock name with
"content-publisher". This could also be done with a WITH_ADVISORY_LOCK_PREFIX
env var, which I think we should set up when we create more of these.
Right now this feels a bit unnecessary as we only have one instance of
usage.

Finally, this has allowed a number of tests to be simplified since the
deleting of preview assets is now part of a background job.