Fix back button bug
As we split the editing image link into two actions here -
https://trello.com/c/TL1An3FR/1077-iterate-crop-edit-workflow-for-lead-and-inline-images
this caused a bug to be introduced where when the user clicks the back
button on the edit image page they are taken back to the crop image
page.

This should only happen when they are uploading a new image, when
they are editing an existing image they should be taken back to the
images index.