Do not fail if content tagged to unknown taxons
Content Publisher only knows about taxons that have the gov.uk homepage
as the root taxon.

However there is another branch of the taxonomy called Worldwide that
has `/world/all` as its root taxon.

`Topic` returns `nil` if is given a taxon content id that it doesn't
recognise.

That means that if a content item is tagged to one of these Worldwide
taxons, Content Publisher throws a 500 exception because it's trying to
populate the miller columns on the edit Topics page with a nil value.

See: https://sentry.io/organizations/govuk/issues/1290560523/?project=1242052&query=is%3Aunresolved+url%3Ahttps%3A%2F%2Fcontent-publisher.publishing.service.gov.uk%2Fdocuments%2Fc12757ba-1093-4eb4-a6a8-15fbcd7588a9%3Aen

This removes the `nil` values from the list of topics, and ignores them
for the purpose of building edit page.

In a later next commit we will make sure that these Worldwide or any
other unknown taxons that aren't in the GOV.UK single taxonomy are not
overwritten when the user updates topic tagging in content publisher.