Retain tagging to taxons not in single taxonomy
Content Publisher only knows about taxons that have the gov.uk homepage
as the root taxon.

However there is another branch of the taxonomy called Worldwide that
has `/world/all` as its root taxon.

This means that when users make changes to tagged taxons in Content
Publisher, they will lose tags added via another app (e.g. Content
Tagger) that does allow content to be tagged to other taxonomies like
the Worldwide one.

Finds the list of taxons that the document is tagged to that Content
Publisher doesn't "know" about, and adds them to the list of taxons that
the user has selected, so that no tags are unexpectedly lost.