Import all users from Whitehall document export
Editors of documents in Whitehall may never have logged into Content
Publisher, so we will not have a row in the users table for them.  This
change creates a user entry for every person who has ever edited the
Whitehall document that is imported.

For security, we give these users no permissions in the Content
Publisher database.  If they log into the application in the future, the
permissions will get updated based on the values in Sign On.

Users who have logged into Content Publisher are skipped.