Raise an exception when Whitehall editions have an unsupported type
There may be documents exported from Whitehall that currently have a
supported document type, but may have historically been a document of a
different type.  This allows us to reject these documents and log an
error.

We are not currently saving the document type to the database, since
this is dependent on work to move document type attribution from the
document to the revision metadata.