Allow document_type_id to be null on the Document table
Trello - https://trello.com/c/kwnmsIe9/1095-permit-content-publisher-to-have-documents-that-change-type

Content publisher currently sets `document_type_id` on the `Document` and `MetadataRevision` table. Before the `document_type_id` column can be removed from `Document` it needs to be able to be set to nill otherwise code changes that do not set it will fail.

This allows the setting of that column to null in order that changes can be made to the underlying code