Import images from Whitehall
Involves creating Image and Image::MetadataRevision in order to
create a valid Image::Revision. Read more about these classes in:
https://github.com/alphagov/content-publisher/blob/master/docs/adr/0009-modelling-history.md

There is no need to create an Image::Asset at this point as it will
be automatically populated when synced with asset manager.

There is no user data associated with images in Whitehall, so we've
had to omit these in places (by explicitly passing `nil`, or by
neglecting to pass a `user` parameter). We could have made an
educated guess by finding the edition associated with the image and
retrieving the last known author for that edition, but there's no
guarantee we'd be picking the right author.