Abort if SVG detected
Decided to repurpose the 'add_filenames_to_images_hash' method
to avoid mapping over the same array twice (more efficient), and
because they're dealing with the same problem domain (extracting
the filename, and analysing the filename to see if it's an SVG).

Unlike other tests in this file, I'm also testing the abort message.
There are an increasing number of reasons why an abort should happen,
and it will provide invaluable debug information to throw a custom
error each time. We should be testing that the error message is what
we expect it to be.