Prevent tags page movement
When JavaScript is enabled the select[multiple] element will be enhanced via JavaScript, but before the enhancement kicks in, the select has the default height which causes quick page movements that are confusing users. We set the select[multiple] element to the default input height of 40px to prevent this.