Decouple UniqueFilenameService from ImageBlobService
Also converts ImageBlobService to hash parameters. I've also reordered
the params so that `temp_image` is the first, as it is the one thing
that is _required_ for the service to work. This work is being done
in advance of importing images from Whitehall.

When we import images from Whitehall using the import script, the way we
use ImageBlobService will have to change:

- There is no 'tempfile.original_filename' method on images downloaded
  from Asset Manager using URI.open.
- There is no 'revision' at the time of calling the ImageBlobService, so
  cannot pass that either.