Add document_type_id to MetadataRevisions
This will allow us to remove it from the Document model which is needed in order to have document types that can change

We decided not to delegate document_id all the way to Document because the document is only meant to hold information that is immutable. We have made these changes with the intention that document_type will become a mutable field.