Potentially create multiple editions for unpublishing
To give some context, an edition in whitehall can be unpublished,
unpublished and edited or unpublished and submitted for 2i all on the
same edition. We needed a way of creating multiple editions if an
unpublished edition has been edited in someway as that's how it should
be represented in content-publisher.

content-publisher refers to a document as removed when it's removed from gov.uk,
whitehall refers to this as an unpublishing.

Some challenges that have occured doing this work:

* An unpublishing can either have a state of draft, submitted or
rejected in whitehall where as in content-publisher this is represented
as a removed edition, or if it's been edited a removed and draft /
submitted_for_2i edition.

* We can't tell what changed when a document is unpublished and edited
in whitehall as all the information is overritten as it's on the same
edition. This means when dividing up an unpublished edition, the only
information we have that's different is the timestamps these events
occured (revision_history).

* Someone in whitehall could unpublish a document and publish it
multiple times which if converted into content-publishers format would
create a load of superseded, unpublished and draft / submitted_for_2i
editions. The current code only creates a maximum of two editions but
this could be the wrong approach if we cared about the full document
representation (including all the superseded editions in
content-publisher). This situation seems unlikely but could happen.

Scenarios that can occur:

* Unpublished edition hasn't been edited or submitted for 2i

We would want to create one edition reflecting the unpublished (removed)
edition.

* Unpublished edition has been edited

We would want to create two editions. One for the unpublishing and one
for the new edited draft.

* Unpublished edition has been submitted_for_2i

We would want to create two editions. One for the unpublishing and one
for the submitted_for_2i. We need to do this as there is no easy way of
knowing if the edition has just been submitted for 2i or if it has been
edited too.

* Unpublished edition has been rejected

We would want to create two editions. One for the unpublishing and one
for the rejected state, in the import script this currently defaults to
submitted_for_2i.