Add created_by user to a status
This adds who created a status for an edition which is the last user to
edit the edition. This could be incorrect however if a status is set by
a user e.g 'submitted' and then is edited by another user. This might be
okay but something to keep in mind.