Remove document_type_id from the Document model
We want to be able to have changing types in content publisher. The Document model is meant to hold immutable information. As such the document type need to be removed from the Document model in order that it can be changed. This removes calls to document_type_id from the Document model.