Test that image type is validated via Requirements::ImageUploadChecker
There is also an explicit check that 'Requirements::ImageUploadChecker' is
called, as otherwise we could relatively easily swap the implementation
out for something that does less validation, and our tests would continue
to pass. We can keep our test suite minimal only because we _know_ we're
delegating that responsibility to Requirements::ImageUploadChecker.