Test that image size is validated via ImageNormaliser
There is an explicit check that 'ImageNormaliser' is called, as
otherwise we could relatively easily swap the implementation out for
something that does less validation, and our tests would continue to
pass. We can keep our test suite minimal only because we _know_ we're
delegating that responsibility to ImageNormaliser.

In testing a 100x100 image, I found that OpenURI treats files differently
if they're smaller than 10KB in size, returning a StringIO instead of
a Tempfile (see https://twin.github.io/improving-open-uri/).
This led to the following error, until I fixed by manually converting
from StringIO to Tempfile:

   # ./lib/tasks/whitehall_importer/create_image_revision.rb:36:in `download_file'
   # ./lib/tasks/whitehall_importer/create_image_revision.rb:13:in `call'
   # ./lib/tasks/whitehall_importer/create_image_revision.rb:5:in `call'
   # ./spec/tasks/whitehall_importer_spec/create_image_revision_spec.rb:80:in `block (4 levels) in <top (required)>'
   # ./spec/tasks/whitehall_importer_spec/create_image_revision_spec.rb:80:in `block (3 levels) in <top (required)>'

We want to validate that the image is large enough to be imported,
but we don't particularly care if an image is _larger_ than 960x640
as it will be automatically cropped.