Import images from Whitehall
This has an integration test / sanity check that the correct number
of instances of classes are created, that they have the right values,
and that the filename auto-increments ('valid-image-1.jpg').
Most of the other tests are in create_image_revision_spec.rb and in
ImageUploadChecker and ImageNormaliser. I felt that it was important
to test that the image filenames are unique as it is
`WhitehallImporter::CreateRevision` that is responsible for passing
the array of 'filenames' to `WhitehallImporter::CreateImageRevision`.