Import governments from Whitehall
This hardcodes government data as config for Content Publisher to have
access to this data. This is intended to be a short-term approach until
governments are exposed to the Publishing API and Content Publisher
consumes them. The effects of duplicating this data is that any changes
to governments will require the config for Content Publisher to be
re-generated.

I decided to pull the data for this directly from the Whitehall database
rather than using the API at https://www.gov.uk/api/governments. The API
seems to only use slugs as identifying information and these slugs seem
to change when governments change to represent the end date, this didn't
strike me as a reliable field we could use and felt the underlying
Whitehall id would be better.