Import a list of political organisations
To determine whether content is political or not for the purposes of
GOV.UK history mode the organisations associated with the content are
considered. A key flag that is checked on these organisations is whether
they are flagged as political, which means that most of their content is
associated with the policies of the current government.

Unfortunately this flag isn't yet exposed to the Publishing API so this
data is locked in Whitehall. In the short term this data is being
hardcoded into Content Publisher until this data can be accessed
dynamically as we race against the date until the 2019 election.