Determine whether edition is political is not
This is a port of the logic from Whitehall that is used to determine
whether an edition is political [1]. This is simplified as Content
Publisher is currently only working with news and press releases and
will need to be expanded with the introduction of more document types
(or logic abstracted to those document types).

Whitehall checks whether content is associated with a minister and a
political organisation. Content Publisher however associates with role
appointments and isn't aware whether these are ministers or not. We
expect that nearly all associations will be ministers so until we have
the ability to determine between these just a check of a role
appointment should suffice.

[1]: https://github.com/alphagov/whitehall/blob/9afd13104412f70d0b739cada49dd1c5629a94b2/lib/political_content_identifier.rb