Change WhitehallImporter to be a module
Rather than act as a class this sets the WhitehallImporter to be a
module. This feels a more semantic way to share the various classes that
share the namespaces.

By wrapping the WhitehallImporter classes in a module we can use this
classes directly rather than having to namespace them as was done when
they were defined under the form of `class
WhitehallImporter::CreateEdition`.

As part of this WhitehallImporter now has a single method of import that
works with the WhitehallImport model, moving this logic from the rake
task. The other logic from WhitehallImporter is now in a Import class.

As part of this each of the classes under WhitehallImporter now have a
consistent `.call` interface.