Switch import specs to use FactoryBot data
This is a rather large change that switches the use of fixture data to
instead use FactoryBot. As part of this there are various tweaks to the
tests, such as adding key ones that were missing, re-ordering and
removing invalid ones.

It also ensures the creation classes return the object they have created
so the return can be tested, with a preference for this over using
`Model.last` approaches.

Apologies that this isn't as granular as it could be :-(