Remove whitehall export fixture data
This has now been replaced with FactoryBot generated data for use in
unit tests.

It wouldn't surprise me too much if we needed to re-introduce fixtures
at some point for integration testing but we were currently only using
this for unit tests and it doesn't represent accurate Whitehall export
data.