Field for a editor determined political status
This field is used to store when an editor has determined that content is
political or not. The value of this field will be used to determine
whether this content enters history mode once the government has
changed.

This field is used to replicate behaviour from Whitehall where an editor
has the ability to override whether an edition is considered political
or not. The possible values for this are:

- null, to indicate it is not being overridden and a system determined political
  status applies;
- true, to indicate an override marking the content as political
- false, to indicate an override marking the content as not political