Add a system_political boolean field to editions
This field is used to store whether Content Publisher has deemed content
to be political or not. It is intended to be populated each time the
edition is synced with the Publishing API and to be paired with the
MetadataRevision#editor_political field to determine political status.

As this is a column with a default value it will lock the db table, I'm
confident that our db is still small enough that this isn't a
performance concern.