Work with quirky aspects of Whitehall Government dates
In Whitehall most governments are specified to end on the date before
the next one starts. However this doesn't seem to be an enforced rule as
the data is inconsistent. For example:

```
{
  "id": "https://www.gov.uk/api/governments/2015-conservative-government",
  "title": "2015 Conservative government",
  "slug": "2015-conservative-government",
  "details": {
    "start_date": "2015-05-08",
    "end_date": null
  }
},
{
  "id": "https://www.gov.uk/api/governments/2010-to-2015-conservative-and-liberal-democrat-coalition-government",
  "title": "2010 to 2015 Conservative and Liberal Democrat coalition government",
  "slug": "2010-to-2015-conservative-and-liberal-democrat-coalition-government",
  "details": {
    "start_date": "2010-05-12",
    "end_date": "2015-05-08"
  }
},
{
  "id": "https://www.gov.uk/api/governments/2005-to-2010-labour-government",
  "title": "2005 to 2010 Labour government",
  "slug": "2005-to-2010-labour-government",
  "details": {
    "start_date": "2005-05-06",
    "end_date": "2010-05-11"
  }
}
```

To cope with these scenarios the #covers method on a Government treats
the end_date as being as if the government continues until 23:59:59 on
that date, so that we don't effectively have no government for 24 hours.

To work with the inconsistency the method to determine a government for
a date works reverse chronologically through the start dates so that
value has a higher precedence than the end date.