Add a column for storing government of an edition
This is used to store the id of the government that content is
associated with. Although this can be worked out by date a static value
is used so that we have a record of whatever was last synced with the
Publishing API and this simplifies looking up whether something is in
history mode or not.