Add publishing_api_payload method to Government
This returns a hash of data that can be included in the Publishing API
details to represent the government. The reflects the data required by
the content-schema for a government [1]. To indicate whether the
government is current has meant the need for also adding the #current?
method.

This approach of including data from another source in the Publishing
API details is somewhat strange and unique to governments. This approach
is intended to go away at the point that governments are represented
with content ids in the Publishing API and instead can use a normal
link. A major downside of this approach is that all content associated
with a current government needs republishing at the point of that
government being replaced.

[1]: https://github.com/alphagov/govuk-content-schemas/blob/68a559e286734464d2578faed5915508c95e3b73/dist/formats/news_article/publisher_v2/schema.json#L277-L299