Add MigrateState class
This class takes some functionality away from CreateStatus which is soon
to be removed. This class was formed as CreateStatus got a bit peculiar
in how it was dealing with status details which was only going to get
more messy with the addition of the unpublishing code soon to come in
later commits.