Stop injecting state into the create event factory
As a create event can only have a state of draft the factory should
reflect this. Changing this only broke one test which required manually
updating the test, if this crops up more we could create a trait for the
whitehall_export/edition_factory that keeps the revision history code
out of the tests itself.