Service to update political fields on Edition
This service can be run with an edition to update the values to indicate
whether the content is political or not. Any content that is changed as
a result of this is marked as not having the revision_synced so that it
is known that the Publishing API needs updating.