Update an editions political association when publishing
This sets the political association at the point of publishing to ensure
politcal status and government are up to date. As we only set government
on published or backdated content we pass in the current government as
something to default to were this to be the first time we are publishing
this content.

In the case that PoliticalAssociationService updates the edition we'll
need to re-sync the preview with the Publishing API so we call the
preview service.