Create EmbedBodyReferences class
Moved the existing 'embed_contacts' method from create_revision.rb
to its own class, which will be responsible for parsing the body
of Whitehall translations and performing regex replacements. In
future commits we will be replacing old Whitehall syntax to standard
Govspeak syntax for embedded images and attachments.

I've added a test to create_revision_spec.rb to ensure that we are
calling the EmbedBodyReferences service so that we only need to have
one test for that. The rest of the body transformation tests will
exist in embed_body_references_spec.rb from now on.