Pass images to EmbedBodyReferences service
Unlike with contacts where we could pass the Whitehall export JSON
directly, we've had to pass the filenames as they appear in the
Image::Revision models themselves. This is because the filenames
have been through parameterisation and deduplication, so there's
no guarantee the filename in the export JSON will continue to match.