Editions aren't created with an association to current government
When this was added in it was expected that all editions would have an
association with a government, with yet to be published content using
the current time to determine the government.

Since then this idea has been revised so that new content is not
associated with a government until the point of publishing (to avoid the
situation where the government can diverge from the content) so this
association is no longer preferable to have by default.