Sync next edition with Publishing API when creating a new edition
When we create a new edition that data should be reflected in the
Publishing API so that there is a draft there to represent the changes.
This is not a hugely consequential change as the changes between draft
and live data is minimal, however this also serves as a point that sets
the political status.

The code to preview the edition used to exist in this class however it
sent the wrong edition to the Publishing API and when we fixed this we
switched to sending no edition to the Publishing API rather than the new
one [1].

[1]: https://github.com/alphagov/content-publisher/commit/f162006f6456dd0bb59e9d7230f9c009b7f08f0c