Flip to use backwards compatible defaults
Rails 6 intoduces some new default values. This turns on all of those that are backwards compatible. Once this is released the `config/initializers/new_framework_defaults_6_0.rb` file can be removed and `config/application.rb` can be updated to use all of the new defaults. This means that we should be able to roll back to rails 5 more easily if needed.