Set an edition link of the government
The Platform Health team have been working to expose governments to the
Publishing API and have now set up the means that content can be linked
to a government. To support this change we are now sending a government
link with our put content.

Until the work to migrate the frontend is complete we will continue to
send government data in the details payload.