Handle withdrawn content
Example withdrawn content:
https://www.gov.uk/api/content/government/news/serious-operating-irregularity-at-romney-sands

Corresponding export:
https://whitehall-admin.integration.publishing.service.gov.uk/government/admin/export/document/426308

It soon transpired that there are numerous states we need to consider,
and rather than handling each case in its own way, we can be a bit
more generic about simply passing an edition to be resynced to the
draft or the live stack.

TBC may need to handle `published_but_needs_2i` state separately.
Not sure the state itself is preserved when publishing.