Build access limit data from just factory methods
The use of the assign_revision and assign_status have led to coupling
that is causing this test to fail when those methods are worked on.
Since we can deal with this at a purely db level we can edit this test
to not be vulnerable to changes to those methods.