Create service to edit edition
This is to replace the use of assign_revision and assign_as_edit in the
edition class. The purpose of this change is to move this business logic
out of the controller into a separate class. This is in preparation of
adding more business logic around editing an edition, which is
anticipated with the automatically determining of government and
political status.

This service intends to update the model without saving it with the
actual saving being the responsibility of the caller. This is so the
method can be used in conjunction with other updates (status updates
often follow a revision update) without saving or at points when not all
data is available to save the edition (such as creating a new edition).