Create service to assign edition statuses
This class pairs with the EditEditionService as the expected common
means to be used to update the status of an edition. It replaces the
assign_status method on the model. This seemed an obvious choice for
removal as calls to it looked different to the EditEditionService calls
and it was one of the larger blocks of code in the edition class.