Set government at the point of publishing
The first time content is published we are likely to not have a
government so we have to set one at the point of publishing. In order to
do this we also need to update the Publishing API if we are to change
the government_id.

We work out the government by checking backdate, the first published and
then current time. Similar logic exists when we associate with a
government when editing an edition, however this does not consider the
current time. As this is slightly different I decided to not move this
into a method - particularly as this logic didn't lend itself to a good
method name.