Fixes following Kevin code review
Create draft of content before marking it as withdrawn.
Avoid using WithdrawService because of its side effects.

Update revision_synced explicitly in the live version
(we're already updating it indirectly in the PreviewService
for draft/current edition).

Reserve base path for both current and live editions in
case the base paths don't match. (Don't anticipate this
for Whitehall content but quite possible for history mode).