Create skeleton ResyncService
This will be responsible for syncing a document with publishing API,
so that we have a reliable way of keeping the two in sync in the
event of a failed import, etc. It will also be used to re-present
content to the publishing API in the event of a change of government,
where some documents will need to be put into 'history mode'. (Ideally
this would happen automatically in publishing API through link
expansion, however governments are not currently represented as links).

I've named it ResyncService for the following reasons:

1) `revision_synced` is a property used throughout the codebase to denote
whether or not the version on content publisher has been synchronised
with the one in publishing API. This keeps our vocabulary consistent.
2) Doesn't necessarily involve publishing (we may just be resyncing a
draft), so `RepublishService` is not appropriate.
3) `RepresentService` would be a viable alternative - but it doesn't
read as "re-present" as we intend it to.